+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release Notes for Invoice Manager
 
+## 2.0.2 - 2025-03-25
+- Handle generator added to invoice template handle field
+- JS Class instantiations moved to twig templates
+
 ## 2.0.1 - 2025-03-24
 - Bug fix: changed Invoice element `tableAttributeHtml()` to `attributeHtml()` for craft 5 compatibility.
-  
+
 ## 2.0.0 - 2025-03-23
 - Craft 5 compatibility